+++ src/Pages/ProductList/ProductSpec/MyBagModal/MyBagModal.scss
@@ -1,0 +1,34 @@
+@import '../../../../Styles/variables.scss';
+
+.myBagModal {
+  position: absolute;
+  background-color: $color-white;
+  top: 0;
+  right: 0;
+  width: 500px;
+  height: 100vh;
+  padding: 20px;
+  transform: translateX(100%);
+  border-left: $border-bottom;
+  border-bottom: $border-bottom;
+  transition: all 0.4s ease-in;
+  z-index: 1;
+
+  .closeBtn {
+    position: absolute;
+    top: 20px;
+    right: 10px;
+    font-size: $font-medium;
+    cursor: pointer;
+  }
+
+  footer {
+    position: absolute;
+    bottom: 10px;
+    width: 90%;
+  }
+
+  &.active {
+    transform: translateX(0);
+  }
+}
