+++ src/Pages/ProductList/ProductSpec/WishListModal/WishListModal.scss
@@ -1,0 +1,48 @@
+@import '../../../../Styles/variables.scss';
+
+.wishListModal {
+  @include flexColumn;
+  align-items: center;
+  position: absolute;
+  top: 70px;
+  right: 70px;
+  width: 380px;
+  height: 500px;
+  padding: 20px;
+  background-color: $color-white;
+  border: $border-bottom;
+  border-radius: $border-radius;
+  z-index: 100;
+
+  .arrow-up {
+    width: 0;
+    height: 0;
+    transform: translateY(-40px);
+    border-left: 20px solid transparent;
+    border-right: 20px solid transparent;
+    border-bottom: 20px solid #dbdbdb;
+  }
+
+  .closeBtn {
+    position: absolute;
+    top: 10px;
+    right: 20px;
+    font-size: $font-medium;
+    cursor: pointer;
+  }
+
+  img {
+    width: 200px;
+    height: 230px;
+  }
+
+  h1 {
+    font-size: $font-medium;
+  }
+
+  p {
+    margin: 20px 0;
+    font-size: 15px;
+    color: $color-gray;
+  }
+}
